too few map points problem tracked to find_orb_correspondence and maybe FeatureMatcher

+++ src/KeyFrame.cpp
@@ -22,7 +22,7 @@ Eigen::Vector3d KeyFrame::fromWorldToImage(Eigen::Vector4d& wcoord) {
 
 float KeyFrame::compute_depth_in_keypoint(cv::KeyPoint kp) {
     // float dd = this->prev_kf->depth_matrix.at<float>(kps[m.queryIdx].pt.y, kps[m.queryIdx].pt.x);
-    uint16_t d = this->depth_matrix.at<uint16_t>(kp.pt.y, kp.pt.x);
+    uint16_t d = this->depth_matrix.at<uint16_t>((int)kp.pt.y, (int)kp.pt.x);
     float dd = d / 5000.0;
     return dd;
 }
